Fixed putnbr

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: epolitze <epolitze@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 14:08:43 by epolitze          #+#    #+#             */
-/*   Updated: 2023/11/21 18:20:05 by epolitze         ###   ########.fr       */
+/*   Updated: 2023/11/22 10:16:14 by epolitze         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,7 @@ int	argument_manager(char c, va_list *arg, int wcount)
 		wcount += temp + ft_putaddress(va_arg(*arg, void *), wcount);
 	}
 	else if (c == 'd' || c == 'i')
-		wcount = ft_putnbr(va_arg(*arg, long long), wcount);
+		wcount = ft_putnbr(va_arg(*arg, int), wcount);
 	else if (c == 'u')
 		wcount = ft_putunbr(va_arg(*arg, unsigned long long), wcount);
 	else if (c == 'x' || c == 'X')
